refactor(editor): useNodeReferences to React context
